+++ database/migrations/2024_05_02_202607_create_barang_masuk_sitohistos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('barang_masuk_sitohistos', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_barang');
-            $table->integer('jumlah_masuk');
+            $table->float('jumlah_masuk');
             $table->date('tanggal_masuk');
             $table->text('keterangan_masuk');
